backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pillow
pyzbar
opencv-python
numpy
requests
PyPDF2
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-multipart
reportlab 
qrcode[pil]
pydantic
Pillow
aiofiles
cloudinary
pytest
pytz